// src/layout/types.ts
export enum AlignItems {
  Start = "flex-start",
  End = "flex-end",
  Center = "center",
  Baseline = "baseline",
  Stretch = "stretch",
}
// ...
export enum Display {
  Block = "block",
  Flex = "flex",
  Initial = "initial",
  Inline = "inline",
  InlineBlock = "inline-block",
  InlineFlex = "inline-flex",
  Hide = "none",
  HideAccessible = "hide-accessible",
}

export type DisplayValue =
  | "block"
  | "flex"
  | "initial"
  | "inline"
  | "inline-block"
  | "inline-flex"
  | "none"
  | "hide-accessible"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum FlexDirection {
  Column = "column",
  ColumnReverse = "column-reverse",
  Row = "row",
  RowReverse = "row-reverse",
}

export type FlexDirectionValue =
  | "column"
  | "column-reverse"
  | "row"
  | "row-reverse"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum FlexWrap {
  Wrap = "wrap",
  NoWrap = "nowrap",
  WrapReverse = "wrap-reverse",
}

export type FlexWrapValue = "wrap" | "nowrap" | "wrap-reverse"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum JustifyContent {
  Start = "flex-start",
  End = "flex-end",
  Center = "center",
  Between = "space-between",
  Around = "space-around",
}
// ...
export enum Position {
  Relative = "relative",
  Absolute = "absolute",
  Fixed = "fixed",
}
// ...
export enum Color {
  Base = "base",
  Alt = "alt",
  Alt2 = "alt-2",
  Link = "link",
  Live = "live",
  Error = "error",
  Overlay = "overlay",
  OverlayAlt = "overlay-alt",
  Inherit = "inherit",
}

export type ColorValue =
  | "base"
  | "alt"
  | "alt-2"
  | "link"
  | "live"
  | "error"
  | "overlay"
  | "overlay-alt"
  | "inherit"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export type Elevation = 0 | 1 | 2 | 3 | 4 | 5;

export enum FontSize {
  Size1 = "font-size-1",
  Size2 = "font-size-2",
  Size3 = "font-size-3",
  Size4 = "font-size-4",
  Size5 = "font-size-5",
  Size6 = "font-size-6",
  Size7 = "font-size-7",
  Size8 = "font-size-8",
}